requirements.txt
# Flask
flask==1.1.2
flask-sqlalchemy==2.4.4
flask-marshmallow==0.14.0
flask-migrate==2.6.0

# Others
gunicorn==20.0.4
pyjwt==2.0.1
python-dotenv==0.15.0
